pyproject.toml
[tool.poetry]
name = "authomate_threat_intelligence"
version = "0.1.0"
description = "A Python package used for ingesting and processing threat intelligence data."
authors = ["Yonah Citron <yonah.citron@shell.com>", "David Schaub <d.schaub@shell.com]"]
packages = [
    { include = "data_accessors", from = "src" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = "3.10.8"
pydantic-settings = "2.3.3"
pydantic = "2.7.3"
pymongo = "4.7.3"
requests = "2.32.3"
pyyaml = "^6.0.1"
azure-functions = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "8.2.2"
pytest-mock = "3.14.0"
pre-commit = "*"
mongomock = "4.1.2"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --cov=data_accessors --cov-report=term-missing"
testpaths = [
    "tests",
]
